voice_agent/requirements.txt
livekit-agents
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
flask
faiss-cpu          
numpy              
PyPDF2       
tqdm
sentence-transformers    
torch  
tiktoken
transformers        
